Extract status/message helpers from taskpane.js into taskpane-status.js
Move 10 pure or near-pure status/message helpers and the
USER_VISIBLE_BANNER_MESSAGE_CODES constant into a new dedicated
module (src/taskpane/taskpane-status.js). All moved helpers are
re-exported and re-imported in taskpane.js so no runtime behavior
changes. Architecture cleanup only.

Moved:
- setStatusMessage
- setCheckMessage
- setInventoryMessage
- nonContiguousSelectionMessage
- buildCheckResolverMessage
- formatBannerUserMessages
- formatBannerUserMessagesExcludingCodes
- formatSelectedRangeGuardrailMessages
- appendSelectedRangeGuardrailMessages
- formatStatusWithSelectedRangeGuardrails
- USER_VISIBLE_BANNER_MESSAGE_CODES (constant)
